refactor(#15): 채팅 목록 가독성 개선
