+++ src/widgets/chat-dialog/ui/tab-contents/friend-section.tsx
@@ -9,34 +9,39 @@ interface FriendSectionProps {
 	onFavoriteToggle?: (friend: FriendInfoResponse) => void;
 }
 
-function FriendSection({ 
-	title, 
-	friends, 
+function FriendSection({
+	title,
+	friends,
 	emptyMessage,
 	onFriendClick,
-	onFavoriteToggle 
+	onFavoriteToggle,
 }: FriendSectionProps) {
+	const renderFriendItems = () => {
+		if (friends.length > 0) {
+			return friends.map((friend) => (
+				<ul key={friend.memberId}>
+					<FriendItem
+						friend={friend}
+						onFriendClick={onFriendClick}
+						onFavoriteToggle={onFavoriteToggle}
+					/>
+				</ul>
+			));
+		}
+
+		if (emptyMessage) {
+			return (
+				<div className="p-4 text-center text-gray-400 text-sm">
+					{emptyMessage}
+				</div>
+			);
+		}
+	};
+
 	return (
 		<div className="mb-6">
 			<h3 className="text-sm regular-11 text-gray-500 mb-3">{title}</h3>
-			<div className="space-y-2">
-				{friends.length > 0 ? (
-					friends.map((friend) => (
-						<FriendItem
-							key={friend.memberId}
-							friend={friend}
-							onFriendClick={onFriendClick}
-							onFavoriteToggle={onFavoriteToggle}
-						/>
-					))
-				) : (
-					emptyMessage && (
-						<div className="p-4 text-center text-gray-400 text-sm">
-							{emptyMessage}
-						</div>
-					)
-				)}
-			</div>
+			<div className="space-y-2">{renderFriendItems()}</div>
 		</div>
 	);
 }
